pyproject.toml
[tool.poetry]
name = "little-leemon-final"
version = "1.0.0"
description = "Terry Brooks, Jr. Final Project forthe APIs course for Meta Backend Developer Certification Via Coursera"
authors = ["Terry Arthur Brooks, Jr. <Terry.Arthur@BrooksJr.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.2"
djangorestframework = "^3.15.2"
Markdown = "^3.7"
django-filter = "^24.3"
djoser = "^2.2.3"
psycopg2-binary = "^2.9.10"
django-cockroachdb = "5.1"
django-seed = "^0.3.1"
djangorestframework-simplejwt = "^5.3.1"
drf-redesign = "^0.4.0"
loguru = "^0.7.2"
django-prometheus = "^2.3.1"
django-cacheops = "^7.1"
django-redis = "^5.4.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
